update(sw): switch Service Worker from `fastest` to `networkFirst` strategy
Switching the Service Worker from `fastest` (cache always wins but at least it gets updated for the next time) to `networkFirst` strategies.
This should still enable offline mode but it should also give you the most up to date website at all times if you are online.
It may have a negative impact on slower devices/networks, but I think we’re OK with that atm.